P5.4: refactor admin page to the TabContent pattern
admin/page.tsx drops from 783 to 124 lines: five tab components in
src/components/admin/, an AdminTable helper for the honestly-shared
table scaffolding, and a useAdminResource fetch hook (with tests) for
self-fetching tabs. The page-level Promise.all fetch triad stays
inline deliberately — its combined loading/error semantics are
behavior. AdminStats joins the api contract types.
